python_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
anndata=0.7.8=pypi_0
argon2-cffi=21.1.0=pypi_0
async-timeout=4.0.1=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39h06a4308_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.7=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pypi_0
commonmark=0.9.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.25=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.0=pypi_0
defusedxml=0.7.1=pypi_0
docrep=0.3.2=pypi_0
entrypoints=0.3=pypi_0
et-xmlfile=1.1.0=pypi_0
falkon=0.6.3=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.28.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2021.11.0=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.41.1=pypi_0
h5py=3.5.0=pypi_0
hyperopt=0.2.6=pypi_0
idna=3.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.5.0=pypi_0
ipython=7.29.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=4.2.1=pypi_0
jupyter-client=7.0.6=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.37.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=5.2.0=pypi_0
natsort=8.0.0=pypi_0
nbclient=0.5.8=pypi_0
nbconvert=6.3.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
notebook=6.4.5=pypi_0
numba=0.54.1=pypi_0
numexpr=2.7.3=pypi_0
numpy=1.20.3=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openpyxl=3.0.9=pypi_0
openssl=1.1.1l=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.2=pypi_0
pandas=1.3.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pypi_0
patsy=0.5.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.22=pypi_0
protobuf=3.19.1=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
py4j=0.10.9.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.0=pypi_0
pygments=2.10.0=pypi_0
pykeops=1.4.2=pypi_0
pynndescent=0.5.5=pypi_0
pyparsing=2.4.7=pypi_0
pyro-api=0.1.2=pypi_0
pyro-ppl=1.7.0=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.3.8=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rich=10.13.0=pypi_0
rsa=4.7.2=pypi_0
scanpy=1.8.2=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.2=pypi_0
scvi-tools=0.14.4=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
setuptools-scm=6.3.2=pypi_0
sinfo=0.3.4=pypi_0
six=1.16.0=pyhd3eb1b0_0
sobolev-alignment=0.0.1=pypi_0
sqlite=3.36.0=hc218d9a_0
statsmodels=0.13.1=pypi_0
stdlib-list=0.8.0=pypi_0
tables=3.6.1=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=1.2.2=pypi_0
torchaudio=0.10.0=py39_cu113
torchmetrics=0.6.0=pypi_0
torchvision=0.11.1=py39_cu113
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
umap-learn=0.5.2=pypi_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0